Move `handleConstraint` to only usage site
